# Conflicts:
#	API/src/main/java/org/sikuli/basics/FileManager.java
#	API/src/main/java/org/sikuli/script/Options.java
#	API/src/main/java/org/sikuli/script/Region.java
#	API/src/main/java/org/sikuli/script/Screen.java
#	API/src/main/java/org/sikuli/script/support/RunTime.java
#	API/src/main/java/org/sikuli/script/support/SikulixAPI.java
#	API/src/main/java/org/sikuli/script/support/SikulixServer.java
#	API/src/main/java/org/sikuli/util/OverlayCapturePrompt.java
#	IDE/src/main/java/org/sikuli/ide/SikuliIDEStatusBar.java
#	IDE/src/main/java/org/sikuli/ide/SikulixIDE.java

+++ IDE/src/main/java/org/sikuli/ide/EditorPane.java
@@ -1061,7 +1061,7 @@ public class EditorPane extends JTextPane {
     return String.format("Region(%d,%d,%d,%d)", x, y, w, h);
   }
 
-  public String getPatternString(String ifn, float sim, Location off, Image img, float resizeFactor, String mask) {
+  public String getPatternString(String ifn, double sim, Location off, Image img, float resizeFactor, String mask) {
 //TODO ifn really needed??
     if (ifn == null) {
       return "\"" + EditorPatternLabel.CAPTURE + "\"";
